+++ CMakeLists.txt
@@ -13,36 +13,7 @@ find_package(sdl2_net CONFIG REQUIRED)
 
 find_path(SDL2_INCLUDE_DIR NAMES SDL.h PATH_SUFFIXES SDL2)
 
-add_executable(butimagine main.cpp World2.cpp util.cpp sockets/sockets.cpp)
-target_include_directories(butimagine PRIVATE ${SDL2_INCLUDE_DIR})
-target_link_libraries(butimagine PRIVATE SDL2::SDL2 SDL2::SDL2main)
-target_link_libraries(butimagine PRIVATE SDL2_image::SDL2_image)
-target_link_libraries(butimagine PRIVATE SDL2_mixer::SDL2_mixer)
-target_link_libraries(butimagine PRIVATE SDL2_ttf::SDL2_ttf)
-target_link_libraries(butimagine PRIVATE SDL2_net::SDL2_net)
-add_compile_options(-DLOG)
-
-add_executable(server sockets/server2.cpp sockets/sockets.cpp World2.cpp util.cpp)
-target_include_directories(server PRIVATE ${SDL2_INCLUDE_DIR})
-target_link_libraries(server PRIVATE SDL2_net::SDL2_net)
-target_link_libraries(server PRIVATE SDL2::SDL2 SDL2::SDL2main)
-target_link_libraries(server PRIVATE SDL2_image::SDL2_image)
-target_link_libraries(server PRIVATE SDL2_mixer::SDL2_mixer)
-target_link_libraries(server PRIVATE SDL2_ttf::SDL2_ttf)
-add_compile_options(-DLOG)
-add_compile_options(-DDEBUG)
-
-add_executable(client sockets/client2.cpp sockets/sockets.cpp World2.cpp util.cpp)
-target_include_directories(client PRIVATE ${SDL2_INCLUDE_DIR})
-target_link_libraries(client PRIVATE SDL2_net::SDL2_net)
-target_link_libraries(client PRIVATE SDL2::SDL2 SDL2::SDL2main)
-target_link_libraries(client PRIVATE SDL2_image::SDL2_image)
-target_link_libraries(client PRIVATE SDL2_mixer::SDL2_mixer)
-target_link_libraries(client PRIVATE SDL2_ttf::SDL2_ttf)
-add_compile_options(-DLOG)
-add_compile_options(-DDEBUG)
-
-add_executable(networking-client sockets/client2.cpp sockets/networking.cpp World2.cpp util.cpp)
+add_executable(networking-client sockets/client.cpp sockets/networking.cpp World.cpp util.cpp)
 target_include_directories(networking-client PRIVATE ${SDL2_INCLUDE_DIR})
 target_link_libraries(networking-client PRIVATE SDL2_net::SDL2_net)
 target_link_libraries(networking-client PRIVATE SDL2::SDL2 SDL2::SDL2main)
@@ -50,8 +21,9 @@ target_link_libraries(networking-client PRIVATE SDL2_image::SDL2_image)
 target_link_libraries(networking-client PRIVATE SDL2_mixer::SDL2_mixer)
 target_link_libraries(networking-client PRIVATE SDL2_ttf::SDL2_ttf)
 add_compile_options(-DLOG)
+add_compile_options(-DDEBUG)
 
-add_executable(networking-server sockets/server2.cpp sockets/networking.cpp World2.cpp util.cpp)
+add_executable(networking-server sockets/server.cpp sockets/networking.cpp World.cpp util.cpp)
 target_include_directories(networking-server PRIVATE ${SDL2_INCLUDE_DIR})
 target_link_libraries(networking-server PRIVATE SDL2_net::SDL2_net)
 target_link_libraries(networking-server PRIVATE SDL2::SDL2 SDL2::SDL2main)
@@ -59,3 +31,4 @@ target_link_libraries(networking-server PRIVATE SDL2_image::SDL2_image)
 target_link_libraries(networking-server PRIVATE SDL2_mixer::SDL2_mixer)
 target_link_libraries(networking-server PRIVATE SDL2_ttf::SDL2_ttf)
 add_compile_options(-DLOG)
+add_compile_options(-DDEBUG)
