+++ sockets/server.cpp
@@ -1,32 +1,86 @@
 #include <cstdio>
-#include <string.h>
-#include <assert.h>
-#include <chrono>
-#include <thread>
-// tutorial maybe https://stephenmeier.net/2015/12/23/sdl-2-0-tutorial-04-networking/
+#include <array>
+#include <windows.h>
+#include <string>
+#include <iostream>
+#include <vector>
 
 #define SDL_MAIN_HANDLED 1
 
 #include "SDL.h"
+#include "SDL_image.h"
+#include "time.h"
+
+#include "../World.hpp"
+
 #include "SDL_net.h"
-#include "sockets.h"
+#include "networking.hpp"
+
+using namespace std;
+
+#pragma warning(disable : 4244)
 
 void init() {
-    if (SDL_Init(SDL_INIT_TIMER|SDL_INIT_EVENTS) != 0) {
-        printf("ER: SDL_Init: %sn", SDL_GetError());
-        assert(false);
-    }
+    IMG_Init(IMG_INIT_JPG | IMG_INIT_PNG);
+    SDL_Init(SDL_INIT_EVERYTHING);
+    //A- Note the global timer starts at SDL initilization
 }
 
 void cleanup() {
     SDL_Quit();
 }
 
-#include <vector>
-int main(int argc, char * argv[]) {
+class ServerDriver {
+    public:
+    Map map;
+    Server ntwk;
+    std::vector<Player*> clientEntities;
+};
+
+ServerDriver driver;
+
+// assumes error handling already done
+void ProcessMoveData(std::vector<Player>& clientPlayers, Packet& packet, int client) {
+    int8_t x = packet.ReadAsType<int8_t>(0);
+    int8_t y = packet.ReadAsType<int8_t>(1);
+    clientPlayers[client].Move(x, y);
+}
+
+void ProcessData(Packet& data, int client) {
+    // if (data.IsMove()) {
+    //     ProcessMoveData(data, client);
+    // }
+}
+
+void SendAllEntities(TCPsocket socket) {
+    // Send them all the entities on the world
+    for (MapEntity* entity : driver.map.allEntities) {
+        Packet newPacket(Packet::Flag_t::bNewEntity);
+        newPacket.EncodePoly(*entity, entity->GetTypeIndex());
+        driver.ntwk.SendPacket(socket, newPacket);
+    }
+    Packet newPacket(Packet::Flag_t::bEndOfPacketGroup);
+    driver.ntwk.SendPacket(socket, newPacket);
+}
+
+void SendPacketToAllButOne(Packet p, int socketIndx) {
+    for (int i = 0; i < driver.ntwk.MAX_SOCKETS; i++) {
+        TCPsocket clientSocket = driver.ntwk.clientSockets[i];
+        if (!clientSocket) {
+            continue;
+        }
+        if (i == socketIndx) {
+            continue;
+        }
+        driver.ntwk.SendPacket(clientSocket, p);
+    }
+}
+
+int main(int argc, char* argv[]) {
+    // process args
     std::vector<std::string> args;
-    if (argc > 2) {
-        for (int i = 0; i < argc - 2; i++) {
+    if (argc > 1) {
+        for (int i = 1; i <= argc - 1; i++) {
             if (!argv[i]) {
                 Log::emit("error reading args to main\n");
             }
@@ -34,107 +88,100 @@ int main(int argc, char * argv[]) {
         }
     }
 
+    const int MAX_PLAYERS = Server::MAX_SOCKETS-1;
+    std::vector<Player> clientPlayers;
 
-    int timeoutTime = 50'000; // in ms, so 50 seconds
-    if (!args.empty()) {
+    int timeoutTime = 500; // in ms
+    if (args.size() != 0) {
         timeoutTime = stoi(args[0]);
     }
     Log::emit("timeout set to %dms\n", timeoutTime);
 
-    // General init (base SDL_init stuff so far)
     init();
-    {
-    // create server socket
-    NetworkHelperServer ntwk;
-    if (!ntwk.Init()) {
-        return -1;
+
+    // Init network connection
+    if (!(driver.ntwk.Setup())) {
+        Log::error("Failed server setup\n");
     }
     
-    int d_sleepsLeft = 100;
-    const int WAIT_TIME = 1000; // ms
     bool runLoop = true;
+    const int WAIT_TIME = 100;
+    while (runLoop) {
 
-    while (runLoop && d_sleepsLeft > 0) {
-        d_sleepsLeft--;
+        if (timeoutTime <= 0) {
+            runLoop = false;
+        }
+        timeoutTime--;
 
-        int clients_ready = SDLNet_CheckSockets(ntwk.socket_set, WAIT_TIME);
+        int clients_ready = SDLNet_CheckSockets(driver.ntwk.socket_set, WAIT_TIME);
 
         if (clients_ready == -1) {
-            printf("Error returned by SDLNet_CheckSockets\n");
+            Log::error("Error returned by SDLNet_CheckSockets\n");
         }
         else if (clients_ready == 0) {
-            printf("No clients ready at this time\n");
+            // Log::emit("No clients ready at this time\n");
             // no clients ready, process things in server work queue
         }
         else if (clients_ready > 0) {
-            // clients are ready
-            printf("Client(s) are ready\n");
-            if(SDLNet_SocketReady(ntwk.serverSoc)) {
-                // client connection ocurred
-                printf("Server socket is ready\n)");
-                ntwk.TryAddClient();
+            // Log::emit("One or more sockets ready\n");
+            if(SDLNet_SocketReady(driver.ntwk.serverSoc)) {
+                // Log::emit("Server socket is ready\n)");
+                int indx = driver.ntwk.TryAddClient();
+                if (indx >= 0) {
+                    Player* joinedPlayer = driver.map.SpawnEntity<Player>();
+                    joinedPlayer->multiplayerID = indx;
+                    joinedPlayer->online = true;
+                    // Send starter data to player client
+                    SendAllEntities(driver.ntwk.clientSockets[joinedPlayer->multiplayerID]);
+                }
             }
-
-            for (int socketIndx = 0; socketIndx < ntwk.MAX_SOCKETS; socketIndx++) {
-                if (!SDLNet_SocketReady(ntwk.clientSockets[socketIndx])) {
+            
+            for (int socketIndx = 0; socketIndx < driver.ntwk.MAX_SOCKETS; socketIndx++) {
+                if (!SDLNet_SocketReady(driver.ntwk.clientSockets[socketIndx])) {
                     continue;
                 }
 
-                printf("Client socket is ready\n)");
+                // Log::emit("Client socket is ready\n)");
 
-                Data data = ntwk.RecvData(ntwk.clientSockets[socketIndx]);
-            
-                // switch(data.dataFlags.bits) {
-                    // case FLAG_WOOD_UPDATE: {
-                    //     Data dataToSend; // TODO: fill this in with real data
-                    //     ntwk.SendData(socketIndx, dataToSend);
-                    // } break;
-            
-                    // case FLAG_QUIT: {
-                    //     runLoop = false;
-                    //     printf("DB: shutdown by client id: %dn", socketIndx);
-                    // } break;
-                //}        
+                Packet packet = driver.ntwk.ConsumePacket(driver.ntwk.clientSockets[socketIndx]);
+
+                // process client packet
+
+                // is it the client moving?
+                if (packet.flags.test(Packet::Flag_t::bMoving)) {
+                    // update our server map
+                    ImMoving moving = packet.ReadAsType<ImMoving>();
+                    Player* client = driver.clientEntities[socketIndx];
+                    client->Move(moving.xOff, moving.yOff);
+
+                    // send notifcation to the other clients about this update
+                    Packet outgoingPacket(Packet::Flag_t::bMoving);
+                    EntityMoveUpdate data = EntityMoveUpdate {client->ID, moving.xOff, moving.yOff};
+                    outgoingPacket.Encode<EntityMoveUpdate>(data);
+                    SendPacketToAllButOne(outgoingPacket, socketIndx);
+                }
             }
         }
-    }
     }
     cleanup();
     return 0;
 }
 
+// stuff I made, and stopped using, but don't wanna throw away yet so I can reference it
+// namespace Junkyard {
+//     void DrawPlayer(Player player, SDL_Surface* winSurface) {
+//         SDL_Rect rect = SDL_Rect {player.position.x, player.position.y, player.width, player.height};
+//         SDL_FillRect( winSurface, &rect, SDL_MapRGB( winSurface->format, 255, 90, 120 ));
+//     }
 
-        // listen on the port for a client, loop until you find one.
-        // I guess there is some way in the socket functions to tell its the client from our game
-        // maybe we can hardcode a passcode in the client, and pass that on its port to this server socket
-
-        // once you have a connection,
-        // wait for it to send you data
-        // probably just keyboard input for now
-
-        // once you have the keyboard input,
-        // we can use the event processing code thats in the eventful-walls branch
-        // + Map, Player, and Wall code in the main branch to move a player on the map, and then return the map back to the client
-
-        // then the client can display the map, and send another keyboard input to the server and repeat
-
-        // server.cpp will have the map, player, and wall objects to keep track of the state of the game
-        // it will also have the logic to parse the keyboard input from the client
-        // then it will send back a the new state of the game for the player to display
-
-        // ORRRRRRR
-
-        // i just got different idea
-        // we could have the map, player, walls, etc on the client side and the server side
-        // so when we do like WASD on the client to move around, we could
-            // 1) process the keyboard input on the client side
-            // 2) update the map with the new player location
-            // 3) send the keyboard input to the server as well
-            // 4) and the server can use the keyboard input to make sure its version of the map
-            //      is updated in the same way
-            // 5) The server will only have to send this data to other clients
-            //      so that they are aware that another client has moved
-            // 
-            // I like this because the size of the data being transferred is a lot smaller
-            // in the first one, its the entire world map, every keystroke
-            // in this one, its just the single keystroke
+//     void InitSurface(Player player1, SDL_Surface* winSurface) {
+//         const int stride = 5;
+//         for (int i = 0; i < MAP_WIDTH; i+=stride) {
+//             for (int j = 0; j < MAP_HEIGHT; j+=stride) {
+//                 SDL_Rect rect {i, j, stride, stride};
+//                 SDL_FillRect(winSurface, &rect, i*MAP_HEIGHT + j);
+//             }
+//         }
+//         DrawPlayer(player1, winSurface);
+//     }
+// };
